--- src/v2/actions/commands.ts
@@ -1,27 +1,0 @@
-export enum ActionCommand {
-	Start = 'start',
-	StartId = 'startid',
-	StartCue = 'startcue',
-	LoadCue = 'loadcue',
-	StartIndex = 'startindex',
-	StartNext = 'start-next',
-	LoadId = 'loadid',
-	LoadIndex = 'loadindex',
-	Pause = 'pause',
-	Stop = 'stop',
-	Reload = 'reload',
-	Next = 'next',
-	Previous = 'previous',
-	Roll = 'roll',
-	Add = 'addtime',
-	SetOnAir = 'set-onair',
-	SetTimerMessageVisibility = 'set-timer-message-visible',
-	SetTimerMessage = 'set-timer-message-text',
-	SetPublicMessageVisibility = 'set-public-message-visible',
-	SetPublicMessage = 'set-public-message-text',
-	SetLowerMessageVisibility = 'set-lower-message-visible',
-	SetLowerMessage = 'set-lower-message-text',
-	SetTimerBlackout = 'set-timer-blackout',
-	SetTimerBlink = 'set-timer-blink',
-	Change = 'change',
-}
