+++ src/v3/ontime-types.ts
@@ -1,4 +1,55 @@
-import { EndAction, Message, OntimeBaseEvent, Playback, SupportedEvent, TimerType } from '../common/ontime-types'
+export enum TimerZone {
+	None = '',
+	Normal = 'normal',
+	Warning = 'warning',
+	Danger = 'danger',
+	Overtime = 'overtime',
+}
+
+//Playbeck
+
+export enum Playback {
+	Roll = 'roll',
+	Play = 'play',
+	Pause = 'pause',
+	Stop = 'stop',
+	Armed = 'armed',
+}
+
+// Message
+
+export type Message = {
+	text: string
+	visible: boolean
+}
+
+// Event
+
+export enum SupportedEvent {
+	Event = 'event',
+	Delay = 'delay',
+	Block = 'block',
+}
+
+export type OntimeBaseEvent = {
+	type: SupportedEvent
+	id: string
+	after?: string // used when creating an event to indicate its position in rundown
+}
+
+export enum EndAction {
+	None = 'none',
+	Stop = 'stop',
+	LoadNext = 'load-next',
+	PlayNext = 'play-next',
+}
+
+export enum TimerType {
+	CountDown = 'count-down',
+	CountUp = 'count-up',
+	TimeToEnd = 'time-to-end',
+	Clock = 'clock',
+}
 
 // Main runtime store
