+++ src/v3/presets.ts
@@ -1,29 +1,143 @@
-import { CompanionButtonPresetDefinition, CompanionPresetDefinitions,  } from '@companion-module/base'
-// import * as icons from '../assets/icons'
-// import { ActionId, feedbackId } from '../enums'
+import {
+	CompanionButtonPresetDefinition,
+	CompanionButtonStyleProps,
+	CompanionPresetDefinitions,
+	combineRgb,
+} from '@companion-module/base'
+import * as icons from '../assets/icons'
+import { ActionId, feedbackId } from '../enums'
+import { TimerZone } from './ontime-types'
 
 export function presets(): CompanionPresetDefinitions {
-	const presets: { [id: string]: CompanionButtonPresetDefinition | undefined } = {}
+	return { ...playbackPresets, ...timerPresets }
+}
 
-	/*
-	presets['start_selected_event'] = {
-		type: 'button',
-		category: 'Commands',
-		name: 'Starts selected event',
+const White = combineRgb(255, 255, 255)
+const Black = combineRgb(0, 0, 0)
+
+const PlaybackGreen = combineRgb(51, 158, 78)
+const PlaybackRed = combineRgb(228, 40, 30)
+const PauseOrange = combineRgb(192, 86, 33)
+const RollBlue = combineRgb(2, 116, 182)
+
+const NormalGray = combineRgb(211, 211, 211)
+const WarningOrange = combineRgb(255, 171, 51)
+const DangerRed = combineRgb(237, 51, 51)
+
+const defaultStyle: CompanionButtonStyleProps = {
+	size: '24',
+	color: White,
+	bgcolor: Black,
+	text: '',
+	alignment: 'center:center',
+	// show_topbar: false,
+}
+
+const defaultWithIconStyle: CompanionButtonStyleProps = {
+	pngalignment: 'center:top',
+	size: '14',
+	color: White,
+	bgcolor: Black,
+	text: '',
+	alignment: 'center:bottom',
+	// show_topbar: false,
+}
+
+const canPlayFeedback = [
+	{
+		feedbackId: feedbackId.ColorPlayback,
+		options: {
+			state: 'play',
+		},
 		style: {
-			png64: icons.PlaybackStart,
-			pngalignment: 'center:top',
-			text: 'START',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(75, 255, 171),
-			bgcolor: combineRgb(0, 0, 0),
+			color: White,
+			bgcolor: PlaybackGreen,
+		},
+	},
+	{
+		feedbackId: feedbackId.ColorPlayback,
+		options: {
+			state: 'armed',
+		},
+		style: {
+			color: PlaybackGreen,
+		},
+	},
+	{
+		feedbackId: feedbackId.ColorPlayback,
+		options: {
+			state: 'pause',
+		},
+		style: {
+			color: PlaybackGreen,
+		},
+	},
+]
+
+const playbackPresets: { [id: string]: CompanionButtonPresetDefinition } = {
+	select_previous_event: {
+		type: 'button',
+		category: 'Playback',
+		name: 'Selects previous event',
+		style: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackPrevious,
 		},
 		steps: [
 			{
 				down: [
 					{
-						actionId: ActionId.Start,
+						actionId: ActionId.Load,
+						options: { method: 'previous' },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: [],
+	},
+	select_next_event: {
+		type: 'button',
+		category: 'Playback',
+		name: 'Selects next event',
+		style: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackNext,
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Load,
+						options: { method: 'next' },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: [],
+	},
+	stop_selected_event: {
+		type: 'button',
+		category: 'Playback',
+		name: 'Stops running event',
+		style: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStop,
+			text: 'STOP',
+			color: PlaybackRed,
+		},
+		previewStyle: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStop,
+			text: 'STOP',
+			bgcolor: PlaybackRed,
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Stop,
 						options: {},
 					},
 				],
@@ -32,27 +146,31 @@ export function presets(): CompanionPresetDefinitions {
 		],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRunning,
-				options: {},
+				feedbackId: feedbackId.ColorPlayback,
+				options: {
+					state: 'stop',
+				},
 				style: {
-					color: combineRgb(255, 255, 255),
-					bgcolor: combineRgb(51, 158, 78),
+					bgcolor: PlaybackRed,
+					color: White,
 				},
 			},
 		],
-	}
-	presets['pause_selected_event'] = {
+	},
+	pause_selected_event: {
 		type: 'button',
-		category: 'Commands',
-		name: 'Pauses selected event',
+		category: 'Playback',
+		name: 'Pauses running event',
 		style: {
+			...defaultWithIconStyle,
 			png64: icons.PlaybackPause,
-			pngalignment: 'center:top',
 			text: 'PAUSE',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(255, 255, 255),
-			bgcolor: combineRgb(0, 0, 0),
+		},
+		previewStyle: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackPause,
+			text: 'PAUSE',
+			bgcolor: PauseOrange,
 		},
 		steps: [
 			{
@@ -67,53 +185,120 @@ export function presets(): CompanionPresetDefinitions {
 		],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorPaused,
-				options: {},
+				feedbackId: feedbackId.ColorPlayback,
+				options: {
+					state: 'pause',
+				},
 				style: {
-					color: combineRgb(255, 255, 255),
-					bgcolor: combineRgb(237, 137, 54),
+					color: White,
+					bgcolor: PauseOrange,
+				},
+			},
+			{
+				feedbackId: feedbackId.ColorPlayback,
+				options: {
+					state: 'play',
+				},
+				style: {
+					color: PauseOrange,
 				},
 			},
 		],
-	}
-	presets['stop_selected_event'] = {
+	},
+	start_selected_event: {
 		type: 'button',
-		category: 'Commands',
-		name: 'Stops selected event',
+		category: 'Playback',
+		name: 'Starts selected event',
 		style: {
-			png64: icons.PlaybackStop,
-			pngalignment: 'center:top',
-			text: 'STOP',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(255, 255, 255),
-			bgcolor: combineRgb(229, 62, 62),
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStart,
+			text: 'START',
+		},
+		previewStyle: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStart,
+			text: 'START',
+			bgcolor: PlaybackGreen,
 		},
 		steps: [
 			{
 				down: [
 					{
-						actionId: ActionId.Stop,
-						options: {},
+						actionId: ActionId.Start,
+						options: {
+							method: 'loaded',
+						},
 					},
 				],
 				up: [],
 			},
 		],
-		feedbacks: [],
-	}
-	presets['reload_selected_event'] = {
+		feedbacks: canPlayFeedback,
+	},
+	start_next_event: {
 		type: 'button',
-		category: 'Commands',
+		category: 'Playback',
+		name: 'Start next event',
+		style: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStart,
+			text: 'NEXT',
+		},
+		previewStyle: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStart,
+			text: 'NEXT',
+			bgcolor: PlaybackGreen,
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Start,
+						options: { method: 'next' },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: canPlayFeedback,
+	},
+	start_selected_or_next_event: {
+		type: 'button',
+		category: 'Playback',
+		name: 'Start selected/next event',
+		style: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStart,
+			text: 'GO',
+		},
+		previewStyle: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackStart,
+			text: 'GO',
+			bgcolor: PlaybackGreen,
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Start,
+						options: { method: 'go' },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: canPlayFeedback,
+	},
+	reload_selected_event: {
+		type: 'button',
+		category: 'Playback',
 		name: 'Reload selected event',
 		style: {
+			...defaultWithIconStyle,
 			png64: icons.PlaybackReload,
-			pngalignment: 'center:top',
 			text: 'RELOAD',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(255, 255, 255),
-			bgcolor: combineRgb(0, 0, 0),
 		},
 		steps: [
 			{
@@ -127,71 +312,22 @@ export function presets(): CompanionPresetDefinitions {
 			},
 		],
 		feedbacks: [],
-	}
-	presets['select_previous_event'] = {
+	},
+	start_roll_mode: {
 		type: 'button',
-		category: 'Commands',
-		name: 'Selects previous event',
-		style: {
-			png64: icons.PlaybackPrevious,
-			pngalignment: 'center:top',
-			text: 'PREVIOUS',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(255, 255, 255),
-			bgcolor: combineRgb(0, 0, 0),
-		},
-		steps: [
-			{
-				down: [
-					{
-						actionId: ActionId.Previous,
-						options: {},
-					},
-				],
-				up: [],
-			},
-		],
-		feedbacks: [],
-	}
-	presets['select_next_event'] = {
-		type: 'button',
-		category: 'Commands',
-		name: 'Selects next event',
-		style: {
-			png64: icons.PlaybackNext,
-			pngalignment: 'center:top',
-			text: 'NEXT',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(255, 255, 255),
-			bgcolor: combineRgb(0, 0, 0),
-		},
-		steps: [
-			{
-				down: [
-					{
-						actionId: ActionId.Next,
-						options: {},
-					},
-				],
-				up: [],
-			},
-		],
-		feedbacks: [],
-	}
-	presets['start_roll_mode'] = {
-		type: 'button',
-		category: 'Commands',
+		category: 'Playback',
 		name: 'Starts Roll Mode',
 		style: {
+			...defaultWithIconStyle,
 			png64: icons.PlaybackRoll,
-			pngalignment: 'center:top',
-			text: 'ROLL MODE',
-			alignment: 'center:bottom',
-			size: '7',
-			color: combineRgb(255, 255, 255),
-			bgcolor: combineRgb(0, 0, 176),
+			text: 'ROLL',
+			color: RollBlue,
+		},
+		previewStyle: {
+			...defaultWithIconStyle,
+			png64: icons.PlaybackRoll,
+			text: 'ROLL',
+			bgcolor: RollBlue,
 		},
 		steps: [
 			{
@@ -206,15 +342,195 @@ export function presets(): CompanionPresetDefinitions {
 		],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRoll,
-				options: {},
+				feedbackId: feedbackId.ColorPlayback,
+				options: { state: 'roll' },
 				style: {
-					color: combineRgb(255, 255, 255),
-					bgcolor: combineRgb(2, 116, 182),
+					color: White,
+					bgcolor: RollBlue,
 				},
 			},
 		],
-	}
+	},
+}
+
+const timerZoneFeedBack = [
+	{
+		feedbackId: feedbackId.TimerZone,
+		options: { zone: TimerZone.Normal },
+		style: { bgcolor: NormalGray, color: Black },
+	},
+	{
+		feedbackId: feedbackId.TimerZone,
+		options: { zone: TimerZone.Warning },
+		style: { bgcolor: WarningOrange, color: Black },
+	},
+	{
+		feedbackId: feedbackId.TimerZone,
+		options: { zone: TimerZone.Danger },
+		style: { bgcolor: DangerRed, color: Black },
+	},
+	{
+		feedbackId: feedbackId.TimerZone,
+		options: { zone: TimerZone.Overtime },
+		style: { bgcolor: Black, color: DangerRed },
+	},
+]
+
+const timerPresets: { [id: string]: CompanionButtonPresetDefinition } = {
+	add_1_min: {
+		type: 'button',
+		category: 'Timer Management',
+		name: 'Add 1 minute to running timer',
+		style: {
+			...defaultStyle,
+			text: '+1',
+			color: PauseOrange,
+			alignment: 'center:center',
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Add,
+						options: { addremove: 'add', minutes: 1, hours: 0, seconds: 0 },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: [],
+	},
+	remove_1_min: {
+		type: 'button',
+		category: 'Timer Management',
+		name: 'Remove 1 minute to running timer',
+		style: { ...defaultStyle, text: '-1', color: PauseOrange, alignment: 'center:center' },
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Add,
+						options: { addremove: 'remove', minutes: 1, hours: 0, seconds: 0 },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: [],
+	},
+	add_5_min: {
+		type: 'button',
+		category: 'Timer Management',
+		name: 'Add 5 minute to running timer',
+		style: {
+			...defaultStyle,
+			text: '+5',
+			color: PauseOrange,
+			alignment: 'center:center',
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Add,
+						options: { addremove: 'add', minutes: 5, hours: 0, seconds: 0 },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: [],
+	},
+	remove_5_min: {
+		type: 'button',
+		category: 'Timer Management',
+		name: 'Remove 5 minute to running timer',
+		style: {
+			...defaultStyle,
+			text: '-5',
+			color: PauseOrange,
+			alignment: 'center:center',
+		},
+		steps: [
+			{
+				down: [
+					{
+						actionId: ActionId.Add,
+						options: { addremove: 'remove', minutes: 5, hours: 0, seconds: 0 },
+					},
+				],
+				up: [],
+			},
+		],
+		feedbacks: [],
+	},
+	current_added: {
+		type: 'button',
+		category: 'Timer Management',
+		name: 'Amount of time added/removed from running timer',
+		style: {
+			...defaultStyle,
+			text: `Total added\n$(ontime:timer_added_nice)`,
+			size: 'auto',
+			alignment: 'center:center',
+		},
+		previewStyle: {
+			...defaultStyle,
+			text: 'Total added\n00',
+			size: 'auto',
+			alignment: 'center:center',
+			bgcolor: PauseOrange,
+		},
+		steps: [
+			{
+				down: [],
+				up: [],
+			},
+		],
+		feedbacks: [
+			{ feedbackId: feedbackId.ColorAddRemove, options: { direction: 'both' }, style: { bgcolor: PauseOrange } },
+		],
+	},
+	current_time_hms: {
+		type: 'button',
+		category: 'Timer Management',
+		name: 'Current timer ',
+		style: {
+			...defaultStyle,
+			text: `$(ontime:time)`,
+			size: '14',
+			alignment: 'center:center',
+		},
+		previewStyle: {
+			...defaultStyle,
+			text: '00:15:25',
+			size: '14',
+			alignment: 'center:center',
+			bgcolor: NormalGray,
+			color: Black,
+		},
+		steps: [
+			{
+				down: [],
+				up: [],
+			},
+		],
+		feedbacks: timerZoneFeedBack,
+	},
+}
+
+/*
+
+
+	//Playback
+
+	
+	
+	
+	
+
+	
+
 	presets['toggle_on_air'] = {
 		type: 'button',
 		category: 'Commands',
@@ -235,7 +551,7 @@ export function presets(): CompanionPresetDefinitions {
 			{
 				down: [
 					{
-						actionId: ActionId.SetOnAir,
+						actionId: deprecatedActionId.SetOnAir,
 						options: { value: 2 },
 					},
 				],
@@ -385,7 +701,7 @@ export function presets(): CompanionPresetDefinitions {
 			{
 				down: [
 					{
-						actionId: ActionId.SetTimerMessageVisibility,
+						actionId: deprecatedActionId.SetTimerMessageVisibility,
 						options: { value: 2 },
 					},
 				],
@@ -394,7 +710,7 @@ export function presets(): CompanionPresetDefinitions {
 		],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.TimerMessageVisible,
+				feedbackId: deprecatedFeedbackId.TimerMessageVisible,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -423,7 +739,7 @@ export function presets(): CompanionPresetDefinitions {
 			{
 				down: [
 					{
-						actionId: ActionId.SetPublicMessageVisibility,
+						actionId: deprecatedActionId.SetPublicMessageVisibility,
 						options: { value: 2 },
 					},
 				],
@@ -432,7 +748,7 @@ export function presets(): CompanionPresetDefinitions {
 		],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.PublicMessageVisible,
+				feedbackId: deprecatedFeedbackId.PublicMessageVisible,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -461,7 +777,7 @@ export function presets(): CompanionPresetDefinitions {
 			{
 				down: [
 					{
-						actionId: ActionId.SetLowerMessageVisibility,
+						actionId: deprecatedActionId.SetLowerMessageVisibility,
 						options: { value: 2 },
 					},
 				],
@@ -470,7 +786,7 @@ export function presets(): CompanionPresetDefinitions {
 		],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.LowerMessageVisible,
+				feedbackId: deprecatedFeedbackId.LowerMessageVisible,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -497,7 +813,7 @@ export function presets(): CompanionPresetDefinitions {
 			{
 				down: [
 					{
-						actionId: ActionId.SetTimerBlink,
+						actionId: deprecatedActionId.SetTimerBlink,
 						options: { value: 2 },
 					},
 				],
@@ -533,7 +849,7 @@ export function presets(): CompanionPresetDefinitions {
 			{
 				down: [
 					{
-						actionId: ActionId.SetTimerBlackout,
+						actionId: deprecatedActionId.SetTimerBlackout,
 						options: { value: 2 },
 					},
 				],
@@ -564,7 +880,7 @@ export function presets(): CompanionPresetDefinitions {
 		steps: [],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRunning,
+				feedbackId: deprecatedFeedbackId.ColorRunning,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 204, 0),
@@ -572,7 +888,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorPaused,
+				feedbackId: deprecatedFeedbackId.ColorPaused,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(237, 137, 54),
@@ -580,7 +896,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorStopped,
+				feedbackId: deprecatedFeedbackId.ColorStopped,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 0, 0),
@@ -588,7 +904,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorRoll,
+				feedbackId: deprecatedFeedbackId.ColorRoll,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -618,7 +934,7 @@ export function presets(): CompanionPresetDefinitions {
 		steps: [],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRunning,
+				feedbackId: deprecatedFeedbackId.ColorRunning,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 204, 0),
@@ -626,7 +942,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorPaused,
+				feedbackId: deprecatedFeedbackId.ColorPaused,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(237, 137, 54),
@@ -634,7 +950,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorStopped,
+				feedbackId: deprecatedFeedbackId.ColorStopped,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 0, 0),
@@ -642,7 +958,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorRoll,
+				feedbackId: deprecatedFeedbackId.ColorRoll,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -672,7 +988,7 @@ export function presets(): CompanionPresetDefinitions {
 		steps: [],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRunning,
+				feedbackId: deprecatedFeedbackId.ColorRunning,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 204, 0),
@@ -680,7 +996,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorPaused,
+				feedbackId: deprecatedFeedbackId.ColorPaused,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(237, 137, 54),
@@ -688,7 +1004,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorStopped,
+				feedbackId: deprecatedFeedbackId.ColorStopped,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 0, 0),
@@ -696,7 +1012,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorRoll,
+				feedbackId: deprecatedFeedbackId.ColorRoll,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -726,7 +1042,7 @@ export function presets(): CompanionPresetDefinitions {
 		steps: [],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRunning,
+				feedbackId: deprecatedFeedbackId.ColorRunning,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 204, 0),
@@ -734,7 +1050,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorPaused,
+				feedbackId: deprecatedFeedbackId.ColorPaused,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(237, 137, 54),
@@ -742,7 +1058,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorStopped,
+				feedbackId: deprecatedFeedbackId.ColorStopped,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 0, 0),
@@ -750,7 +1066,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorRoll,
+				feedbackId: deprecatedFeedbackId.ColorRoll,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -780,7 +1096,7 @@ export function presets(): CompanionPresetDefinitions {
 		steps: [],
 		feedbacks: [
 			{
-				feedbackId: feedbackId.ColorRunning,
+				feedbackId: deprecatedFeedbackId.ColorRunning,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 204, 0),
@@ -788,7 +1104,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorPaused,
+				feedbackId: deprecatedFeedbackId.ColorPaused,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(237, 137, 54),
@@ -796,7 +1112,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorStopped,
+				feedbackId: deprecatedFeedbackId.ColorStopped,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(0, 0, 0),
@@ -804,7 +1120,7 @@ export function presets(): CompanionPresetDefinitions {
 				options: {},
 			},
 			{
-				feedbackId: feedbackId.ColorRoll,
+				feedbackId: deprecatedFeedbackId.ColorRoll,
 				style: {
 					color: combineRgb(255, 255, 255),
 					bgcolor: combineRgb(43, 108, 176),
@@ -834,6 +1150,5 @@ export function presets(): CompanionPresetDefinitions {
 		steps: [],
 		feedbacks: [],
 	}
-	*/
-	return presets
-}
+
+*/
